Improve gradle performance significantly
