+++ buildSrc/src/main/kotlin/com/dfsek/terra/DependencyConfig.kt
@@ -2,12 +2,24 @@ package com.dfsek.terra
 
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.invoke
 import org.gradle.kotlin.dsl.repositories
 
 fun Project.configureDependencies() {
 
+    configurations {
+        val shaded = create("shaded")
+        getByName("compile").extendsFrom(shaded)
+        val shadedApi = create("shadedApi")
+        shaded.extendsFrom(shadedApi)
+        getByName("api").extendsFrom(shadedApi)
+        val shadedImplementation = create("shadedImplementation")
+        shaded.extendsFrom(shadedImplementation)
+        getByName("implementation").extendsFrom(shadedImplementation)
+    }
+
     repositories {
-        maven { url = uri("http://maven.enginehub.org/repo/") }
+        maven { url = uri("https://maven.enginehub.org/repo/") }
         maven { url = uri("https://repo.codemc.org/repository/maven-public") }
         maven { url = uri("https://papermc.io/repo/repository/maven-public/") }
         maven { url = uri("https://maven.fabricmc.net/") }
@@ -19,6 +31,6 @@ fun Project.configureDependencies() {
     dependencies {
         "testImplementation"("org.junit.jupiter:junit-jupiter-api:5.7.0")
         "testImplementation"("org.junit.jupiter:junit-jupiter-engine:5.7.0")
-        "compileOnly"("org.jetbrains:annotations:20.1.0")
+        "api"("org.jetbrains:annotations:20.1.0")
     }
 }
