+++ common/build.gradle.kts
@@ -1,11 +1,15 @@
 import com.dfsek.terra.configureCommon
+import com.dfsek.terra.configureCompilation
+import com.dfsek.terra.configureDependencies
 
 plugins {
     `java-library`
     `maven-publish`
+    idea
 }
 
-configureCommon()
+configureCompilation()
+configureDependencies()
 
 group = "com.dfsek.terra.common"
