+++ gradle.properties
@@ -1,0 +1,10 @@
+# Magic options for more perf
+org.gradle.jvmargs=-Xmx2048M -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8 -XX:+UseParallelGC
+org.gradle.vfs.watch=true
+kapt.use.worker.api=true
+kapt.include.compile.classpath=false
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.warning.mode=all
+#org.gradle.logging.level=info
+org.gradle.configureondemand=true
